+++ lista1/src/tsp/geo.rs
@@ -1,37 +1,37 @@
 #[derive(Debug)]
 pub struct Point {
-    pub x: f64,
-    pub y: f64,
+    pub x: i64,
+    pub y: i64,
 }
 
 #[allow(dead_code)]
 impl Point {
-    pub fn distance(&self, another: &Point) -> f64 {
-        let x2 = (self.x - another.x).powf(2.0);
-        let y2 = (self.y - another.y).powf(2.0);
-        return (x2 + y2).sqrt();
+    pub fn distance(&self, another: &Point) -> u64 {
+        let x2 = (self.x - another.x).pow(2) as f64;
+        let y2 = (self.y - another.y).pow(2) as f64;
+        return (x2 + y2).sqrt() as u64;
     }
 }
 
 #[derive(Debug)]
 pub struct Matrix {
     pub n: usize,
-    pub matrix: Vec<f64>,
+    pub matrix: Vec<u64>,
 }
 
 
 #[allow(dead_code)]
 impl Matrix {
-    pub fn get(&self, x: usize, y: usize) -> f64 {
+    pub fn get(&self, x: usize, y: usize) -> u64 {
         return self.matrix[x * self.n + y];
     }
 
-    pub fn put(&mut self, x: usize, y: usize, value: f64) {
+    pub fn put(&mut self, x: usize, y: usize, value: u64) {
         self.matrix[x * self.n + y] = value;
     }
 
     pub fn new(n: usize) -> Matrix {
-        let matrix = vec![0.0; n * n];
+        let matrix = vec![0; n * n];
         return Matrix { n, matrix }
     }
 }
